requirements.txt
-i https://pypi.org/simple
blinker==1.9.0; python_version >= '3.9'
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
dash==3.0.2; python_version >= '3.8'
dash-auth==2.3.0; python_version >= '3.8'
flask==3.0.3; python_version >= '3.8'
gunicorn==23.0.0; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
importlib-metadata==8.6.1; python_version >= '3.9'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
narwhals==1.34.1; python_version >= '3.8'
nest-asyncio==1.6.0; python_version >= '3.5'
numpy==2.2.4; python_version >= '3.10'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
plotly==6.0.1; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.1.0; python_version >= '3.9'
pytz==2025.2
requests==2.32.3; python_version >= '3.8'
retrying==1.3.4
setuptools==78.1.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
typing-extensions==4.13.2; python_version >= '3.8'
tzdata==2025.2; python_version >= '2'
urllib3==2.4.0; python_version >= '3.9'
werkzeug==3.0.6; python_version >= '3.8'
zipp==3.21.0; python_version >= '3.9'
